+++ common/src/main/java/mca/advancement/criterion/RankCriterion.java
@@ -2,6 +2,7 @@ package mca.advancement.criterion;
 
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
+import mca.MCA;
 import mca.resources.Rank;
 import net.minecraft.advancement.criterion.AbstractCriterion;
 import net.minecraft.advancement.criterion.AbstractCriterionConditions;
@@ -12,7 +13,7 @@ import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
 
 public class RankCriterion extends AbstractCriterion<RankCriterion.Conditions> {
-    private static final Identifier ID = new Identifier("mca:rank");
+    private static final Identifier ID = MCA.locate("rank");
 
     @Override
     public Identifier getId() {