+++ common/src/main/java/mca/resources/ApiReloadListener.java
@@ -1,11 +1,12 @@
 package mca.resources;
 
+import mca.MCA;
 import net.minecraft.resource.ResourceManager;
 import net.minecraft.resource.SynchronousResourceReloader;
 import net.minecraft.util.Identifier;
 
 public class ApiReloadListener implements SynchronousResourceReloader {
-    public static final Identifier ID = new Identifier("mca", "api");
+    public static final Identifier ID = MCA.locate("api");
 
     @Override
     public void reload(ResourceManager manager) {