+++ common/src/main/java/mca/server/world/data/PlayerSaveData.java
@@ -34,6 +34,7 @@ import net.minecraft.text.TranslatableText;
 import net.minecraft.util.Formatting;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.PersistentState;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
@@ -65,6 +66,7 @@ public class PlayerSaveData extends PersistentState implements EntityRelationshi
         assert playerId != null;
         this.world = world;
         this.playerId = playerId;
+        this.marriageState = MarriageState.SINGLE;
         resetEntityData();
     }
 
@@ -72,9 +74,10 @@ public class PlayerSaveData extends PersistentState implements EntityRelationshi
         this.world = world;
         playerId = nbt.getUuid("playerId");
         lastSeenVillage = nbt.contains("lastSeenVillage", NbtElement.INT_TYPE) ? Optional.of(nbt.getInt("lastSeenVillage")) : Optional.empty();
-        spouseUUID = nbt.contains("spouseUUID", NbtElement.INT_TYPE) ? Optional.of(nbt.getUuid("spouseUUID")) : Optional.empty();
+        spouseUUID = nbt.contains("spouseUUID") ? Optional.of(nbt.getUuid("spouseUUID")) : Optional.empty();
         spouseName = nbt.contains("spouseName") ? Optional.of(new LiteralText(nbt.getString("spouseName"))) : Optional.empty();
         entityDataSet = nbt.contains("entityDataSet") && nbt.getBoolean("entityDataSet");
+        marriageState = MarriageState.byId(nbt.getInt("marriageState"));
 
         if (nbt.contains("entityData")) {
             entityData = nbt.getCompound("entityData");
@@ -234,8 +237,8 @@ public class PlayerSaveData extends PersistentState implements EntityRelationshi
     }
 
     @Override
-    public FamilyTreeNode getFamilyEntry() {
-        return getFamilyTree().getOrCreate(world.getEntity(playerId));
+    public @NotNull FamilyTreeNode getFamilyEntry() {
+        return getFamilyTree().getOrCreate(Objects.requireNonNull(world.getEntity(playerId)));
     }
 
     @Override
@@ -272,6 +275,7 @@ public class PlayerSaveData extends PersistentState implements EntityRelationshi
         nbt.put("entityData", entityData);
         nbt.putBoolean("entityDataSet", entityDataSet);
         nbt.put("inbox", NbtHelper.fromList(inbox, v -> v));
+        nbt.putInt("marriageState", marriageState.ordinal());
         return nbt;
     }
 
@@ -301,7 +305,7 @@ public class PlayerSaveData extends PersistentState implements EntityRelationshi
                 getFamilyEntry().getName(), name, village)));
         NbtCompound nbt = new NbtCompound();
         nbt.put("pages", l);
-        inbox.add(nbt);
+        sendMail(nbt);
         showMailNotification(player);
     }